requirement.txt
flask
langchain-openai
langchain_mistralai
flask_socketio